# docs/swagger.yaml
basePath: /
definitions:
  authentication.AuthLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.CreateUserDTO:
    properties:
      email:
        type: string
      name:
        minLength: 5
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: Example API with Gin, GORM, Clean Architecture
  title: MyApp API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authentication.AuthLoginDTO'
      produces:
      - application/json
      responses: {}
      summary: Login user
      tags:
      - Authentication
  /profile:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get profile
      tags:
      - profile
  /users:
    get:
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a user
      parameters:
      - description: User body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDTO'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get user by id
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"